Rename spacy.analysis to spacy.pipe_analysis
